pyproject.toml
[tool.poetry]
name = "e-commerce"
version = "0.2.0"
description = "Final project for the MLOps course"
authors = ["Juan Sebastián Ramírez Artiles <juanseraar@hotmail.com>"]
readme = "README.md"
packages = [{include = "e_commerce"}]
include = ["configuration.yaml"]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^1.5.2"
pyarrow = "^10.0.1"
typer = "^0.7.0"
scikit-learn = "^1.1.3"
pyyaml = "^6.0"


[tool.poetry.group.dev.dependencies]
mypy = "^0.991"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
